requirements.txt
alembic==1.2.1
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
